Verify is package is clean before deleting

+++ src/commands/packages/rm.command.ts
@@ -1,6 +1,8 @@
 import {Argv, Arguments} from "yargs";
+const { prompt } = require('enquirer');
 
 import runScript from '../../tools/run-script';
+import runCommand from '../../tools/run-command';
 import assertProject from '../../tools/assert-project';
 
 import { Global } from '../../constants/types';
@@ -25,6 +27,22 @@ export const handler = async (argv : Arguments<HandlerArguments>) => {
   const { name } = argv;
 
   assertProject();
+
+  const { stdout : modified } = await runCommand(`git status packages/${name} --porcelain`, false, {cwd : global.PROJECT_ROOT});
+
+  if (modified) {
+    console.error(`ERROR : Submodule ${name} was modified`)
+    process.exitCode = 1;
+    return;
+  }
+
+  const { removePackage } = await prompt({
+    type: 'confirm',
+    name: 'removePackage',
+    message: `Reference to package ${name} will be removed from project. Continue?`
+  })
+
+  if (!removePackage) {return;}
 
   runScript(`
     git rm packages/${name}
